+++ src/apis/auth/token.ts
@@ -8,8 +8,10 @@ const TOKEN_KEYS = {
 
 // 초기화: localStorage에서 토큰 복원
 const initializeTokens = () => {
-  accessToken = localStorage.getItem(TOKEN_KEYS.ACCESS);
-  refreshToken = localStorage.getItem(TOKEN_KEYS.REFRESH);
+  if (typeof window !== 'undefined') {
+    accessToken = localStorage.getItem(TOKEN_KEYS.ACCESS);
+    refreshToken = localStorage.getItem(TOKEN_KEYS.REFRESH);
+  }
 };
 
 initializeTokens();
@@ -21,14 +23,18 @@ export const token = {
   setTokens: (access: string, refresh: string) => {
     accessToken = access;
     refreshToken = refresh;
-    localStorage.setItem(TOKEN_KEYS.ACCESS, access);
-    localStorage.setItem(TOKEN_KEYS.REFRESH, refresh);
+    if (typeof window !== 'undefined') {
+      localStorage.setItem(TOKEN_KEYS.ACCESS, access);
+      localStorage.setItem(TOKEN_KEYS.REFRESH, refresh);
+    }
   },
 
   clearTokens: () => {
     accessToken = null;
     refreshToken = null;
-    localStorage.removeItem(TOKEN_KEYS.ACCESS);
-    localStorage.removeItem(TOKEN_KEYS.REFRESH);
+    if (typeof window !== 'undefined') {
+      localStorage.removeItem(TOKEN_KEYS.ACCESS);
+      localStorage.removeItem(TOKEN_KEYS.REFRESH);
+    }
   },
 };
