--- src/components/common/Header/types.ts
@@ -1,8 +1,0 @@
-export interface Notification {
-  id: number;
-  status: 'approved' | 'rejected';
-  title: string;
-  reservationTime: string;
-  time: string;
-  createdAt: string;
-}
