+++ src/utils/timeUtils.ts
@@ -1,0 +1,23 @@
+// 시간 표시 포맷팅 함수
+export const getTimeAgo = (createdAt: string) => {
+  const now = new Date();
+  const created = new Date(createdAt);
+  const diffMs = now.getTime() - created.getTime();
+  const diffMins = Math.floor(diffMs / 60000);
+  const diffHours = Math.floor(diffMins / 60);
+  const diffDays = Math.floor(diffHours / 24);
+
+  if (diffMins < 1) {
+    return '방금 전';
+  }
+  if (diffMins < 60) {
+    return `${diffMins}분 전`;
+  }
+  if (diffHours < 24) {
+    return `${diffHours}시간 전`;
+  }
+  if (diffDays < 7) {
+    return `${diffDays}일 전`;
+  }
+  return created.toLocaleDateString('ko-KR');
+};
